backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
distro==1.9.0
dj-database-url==3.1.2
Django==6.0.3
django-cors-headers==4.9.0
django-sendgrid-v5==1.3.1
djangorestframework==3.16.1
drf-spectacular==0.29.0
fastjsonschema==2.21.2
fonttools==4.61.1
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.191.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
graphviz==0.21
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
inflection==0.5.1
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
kiwisolver==1.4.9
kombu==5.6.2
MarkupSafe==3.0.3
matplotlib==3.10.8
nbformat==5.10.4
numpy==2.4.2
openai==2.24.0
packaging==26.0
panda==0.3.1
pandas==3.0.1
pillow==12.1.1
platformdirs==4.9.2
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==5.29.6
psycopg2==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-http-client==3.3.7
PyYAML==6.0.3
redis==7.2.1
referencing==0.37.0
reportlab==4.4.10
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.1
sendgrid==6.12.5
setuptools==82.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
threadpoolctl==3.6.0
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
Werkzeug==3.1.6
